Refactored of code in all adapters and activities

+++ app/src/main/java/com/applocum/fitzoh/ui/profile/activities/ProfileFragment.kt
@@ -1,6 +1,7 @@
 package com.applocum.fitzoh.ui.profile.activities
 
 import android.content.Context.MODE_PRIVATE
+import android.content.DialogInterface
 import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
@@ -8,10 +9,13 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import com.applocum.fitzoh.Dbhelper
 import com.applocum.fitzoh.R
+import com.applocum.fitzoh.ui.home.activities.HolisticHealthTransferActivity
 import com.applocum.fitzoh.ui.mainscreen.activities.MainActivity
+import kotlinx.android.synthetic.main.fragment_profile.*
 import kotlinx.android.synthetic.main.fragment_profile.view.*
 
 class ProfileFragment : Fragment() {
@@ -24,16 +28,25 @@ class ProfileFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         val v=inflater.inflate(R.layout.fragment_profile, container, false)
-        v.tvLogout.setOnClickListener {
-            sharedPreferences=activity!!.getSharedPreferences("mypref",MODE_PRIVATE)
-            editor=sharedPreferences.edit()
-            editor.clear()
-            editor.apply()
+        v.rlsignout.setOnClickListener {
+            activity?.let { it1 ->
+                AlertDialog.Builder(it1)
+                    .setTitle("Sign out?")
+                    .setMessage("Are you sure you want to Sign Out?")
+                    .setNegativeButton("Cancel", null)
+                    .setPositiveButton("Sign out", DialogInterface.OnClickListener { dialog, which ->
+                        sharedPreferences=requireActivity().getSharedPreferences("mypref",MODE_PRIVATE)
+                        editor=sharedPreferences.edit()
+                        editor.clear()
+                        editor.apply()
 
-            val intent= Intent(activity,
-                MainActivity::class.java)
-            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
-            startActivity(intent)
+                        val intent= Intent(activity,
+                            MainActivity::class.java)
+                        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
+                        startActivity(intent)
+
+                    }).create().show()
+            }
         }
 
         v.tvEditProfile.setOnClickListener {
@@ -41,17 +54,34 @@ class ProfileFragment : Fragment() {
             startActivity(intent)
         }
 
+        v.rlPurchase.setOnClickListener {
+            val intent=Intent(activity,HolisticHealthTransferActivity::class.java)
+            startActivity(intent)
+        }
 
-        val dbhelper= activity?.let { Dbhelper(it) }
+      /*  val dbhelper= activity?.let { Dbhelper(it) }
 
-        sharedPreferences=activity!!.getSharedPreferences("mypref", AppCompatActivity.MODE_PRIVATE)
+        sharedPreferences=requireActivity().getSharedPreferences("mypref", AppCompatActivity.MODE_PRIVATE)
         val email=sharedPreferences.getString("email","")
 
        val user= email?.let { dbhelper?.signin(it) }
 
         v.tvKg.text = user?.userWeight
-        v.tvName.text = user?.userName
+        v.tvName.text = user?.userName*/
         return v
+
+    }
+
+    override fun onResume() {
+        super.onResume()
+        val dbhelper= activity?.let { Dbhelper(it) }
+
+        sharedPreferences=requireActivity().getSharedPreferences("mypref", AppCompatActivity.MODE_PRIVATE)
+        val email=sharedPreferences.getString("email","")
+        val user= email?.let { dbhelper?.signin(it) }
+
+        tvKg.text = user?.userWeight
+        tvName.text = user?.userName
 
     }
 }
